finish refactoring the code
