+++ Cargo.toml
@@ -24,4 +24,5 @@ serde = "^1.0.203"
 serde_json = "^1.0.117"
 sqlx = { version = "^0.7", features = ["runtime-tokio-rustls", "any"] }
 tokio = { version = "^1.38.0", features = ["macros", "rt-multi-thread"] }
-rusql-alchemy-macro = "0.1.1"
+# rusql-alchemy-macro = "0.1.1"
+rusql-alchemy-macro = { version = "*", git = "https://github.com/russenger/rusql-alchemy-macro", branch = "main" }
