+++ examples/postgres/.env
@@ -1,1 +1,2 @@
-DATABASE_URL=postgres://<username>:<password>@<hostname>/<dbname>
+# DATABASE_URL=postgres://<username>:<password>@<hostname>/<dbname>
+DATABASE_URL=postgres://test:test@localhost/test
