+++ examples/sqlite/.env
@@ -1,1 +1,2 @@
-DATABASE_URL=sqlite://<dbname>
+# DATABASE_URL=sqlite://<dbname>
+DATABASE_URL=sqlite://test.db
