+++ src/prelude.rs
@@ -1,11 +1,10 @@
 #[cfg(feature = "postgres")]
-pub use crate::db::models::Serial;
+pub use super::types::Serial;
 
-pub use crate::Connection;
-pub use crate::{
-    db::{models::*, Database},
-    kwargs, migrate,
-};
+pub use super::types::*;
+pub use super::Connection;
+pub use super::Database;
+pub use super::{db::models::*, kwargs, migrate};
 pub use async_trait::async_trait;
 pub use rusql_alchemy_macro::Model;
 pub use sqlx::FromRow;
